Revise TrackSchemeZoom.
